# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Couples shared planning app API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CoupleProfile {
  id: number;
  partner1Name: string;
  partner2Name: string;
  partner1Status: string;
  partner2Status: string;
  /** @nullable */
  partner1Emoji?: string | null;
  /** @nullable */
  partner2Emoji?: string | null;
  /** @nullable */
  anniversaryDate: string | null;
  /** @nullable */
  coupleTitle?: string | null;
  createdAt?: string;
}

export interface CoupleProfileUpdate {
  partner1Name?: string;
  partner2Name?: string;
  partner1Emoji?: string;
  partner2Emoji?: string;
  anniversaryDate?: string;
  coupleTitle?: string;
}

export interface StatusUpdate {
  status: string;
  emoji?: string;
}

export interface Todo {
  id: number;
  title: string;
  completed: boolean;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  addedBy?: string | null;
  createdAt: string;
}

export interface TodoInput {
  /** @minLength 1 */
  title: string;
  note?: string;
  addedBy?: string;
}

export interface TodoUpdate {
  /** @minLength 1 */
  title?: string;
  completed?: boolean;
  note?: string;
  addedBy?: string;
}

export interface Goal {
  id: number;
  title: string;
  completed: boolean;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  targetDate?: string | null;
  /** @nullable */
  category?: string | null;
  createdAt: string;
}

export interface GoalInput {
  /** @minLength 1 */
  title: string;
  note?: string;
  targetDate?: string;
  category?: string;
}

export interface GoalUpdate {
  title?: string;
  completed?: boolean;
  note?: string;
  targetDate?: string;
  category?: string;
}

export interface WishItem {
  id: number;
  title: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  link?: string | null;
  /** @nullable */
  addedBy?: string | null;
  purchased?: boolean;
  createdAt: string;
}

export interface WishItemInput {
  /** @minLength 1 */
  title: string;
  note?: string;
  link?: string;
  addedBy?: string;
}

export interface WishItemUpdate {
  title?: string;
  note?: string;
  link?: string;
  addedBy?: string;
  purchased?: boolean;
}

export interface BucketItem {
  id: number;
  title: string;
  completed: boolean;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  targetDate?: string | null;
  createdAt: string;
}

export interface BucketItemInput {
  /** @minLength 1 */
  title: string;
  note?: string;
  targetDate?: string;
}

export interface BucketItemUpdate {
  title?: string;
  completed?: boolean;
  note?: string;
  targetDate?: string;
}

export interface CalendarEvent {
  id: number;
  title: string;
  date: string;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  emoji?: string | null;
  isAnniversary?: boolean;
  createdAt: string;
}

export interface CalendarEventInput {
  /** @minLength 1 */
  title: string;
  date: string;
  note?: string;
  emoji?: string;
  isAnniversary?: boolean;
}

export interface CalendarEventUpdate {
  title?: string;
  date?: string;
  note?: string;
  emoji?: string;
  isAnniversary?: boolean;
}

export interface DashboardSummary {
  todoCount: number;
  todoDoneCount: number;
  goalCount: number;
  goalDoneCount: number;
  wishlistCount: number;
  bucketCount: number;
  bucketDoneCount: number;
  upcomingEvents: CalendarEvent[];
}

export type ListEventsParams = {
month?: number;
year?: number;
};