# lib/api-zod/src/generated/types/bucketItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Couples shared planning app API
 * OpenAPI spec version: 0.1.0
 */

export interface BucketItemInput {
  /** @minLength 1 */
  title: string;
  note?: string;
  targetDate?: string;
}