# lib/api-zod/src/generated/types/bucketItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Couples shared planning app API
 * OpenAPI spec version: 0.1.0
 */

export interface BucketItemUpdate {
  title?: string;
  completed?: boolean;
  note?: string;
  targetDate?: string;
}